Binh

+++ src/main/java/com/example/demo/service/ChiTietSanPhamService.java
@@ -3,6 +3,8 @@ package com.example.demo.service;
 
 import com.example.demo.entity.ChiTietSanPham;
 import com.example.demo.repository.ChiTietSanPhamRepository;
+import com.example.demo.repository.GioHangChiTietRepository;
+import com.example.demo.repository.HinhAnhRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -15,6 +17,10 @@ public class ChiTietSanPhamService {
 
     @Autowired
     private ChiTietSanPhamRepository chiTietSanPhamRepository;
+    @Autowired
+    private HinhAnhRepository hinhAnhRepository;
+    @Autowired
+    private GioHangChiTietRepository gioHangChiTietRepository;
 
     public List<ChiTietSanPham> getAll() {
         return chiTietSanPhamRepository.findAll();
@@ -82,7 +88,13 @@ public class ChiTietSanPhamService {
     }
 
     public void deleteChiTietSanPham(Integer id) {
-        chiTietSanPhamRepository.deleteById(id);
+        if(gioHangChiTietRepository.countByChiTietSanPhamId(id)==0||hinhAnhRepository.coundChitiet(id)==0){
+            chiTietSanPhamRepository.deleteById(id);
+        }else {
+            ChiTietSanPham existingChiTietSanPham = chiTietSanPhamRepository.findById(id).orElse(null);
+            existingChiTietSanPham.setTrangThai(false);
+            chiTietSanPhamRepository.save(existingChiTietSanPham);
+        }
     }
 }
